loadbalancer: Move pkg/service to pkg/loadbalancer/legacy/service
Relocate the service package to pkg/loadbalancer/legacy/service to
clearly mark it as deprecated and on its way out.
