+++ pkg/ciliumenvoyconfig/legacy/envoy_l7lb_backend_syncer.go
@@ -14,9 +14,9 @@ import (
 
 	"github.com/cilium/cilium/pkg/envoy"
 	"github.com/cilium/cilium/pkg/loadbalancer"
+	"github.com/cilium/cilium/pkg/loadbalancer/legacy/service"
 	"github.com/cilium/cilium/pkg/lock"
 	"github.com/cilium/cilium/pkg/logging/logfields"
-	"github.com/cilium/cilium/pkg/service"
 	"github.com/cilium/cilium/pkg/slices"
 )
